Check existence

+++ install.sh
@@ -3,5 +3,7 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 ENTITIES=(.vim .vimrc .gitignore_global .gitconfig)
 for entity in ${ENTITIES[@]}; do
-  ln -s "$DIR/$entity" "$HOME/$entity"
+  if [ ! -L "$HOME/$entity" ]; then
+    ln -s "$DIR/$entity" "$HOME/$entity"
+  fi
 done
